+++ js/utils.js
@@ -1,0 +1,19 @@
+// eslint-disable-next-line id-length
+function getRandomFloat (a, b, digits = 1) {
+  const lower = Math.min(Math.abs(a), Math.abs(b));
+  const upper = Math.max(Math.abs(a), Math.abs(b));
+  const result = Math.random() * (upper - lower) + lower;
+  return result.toFixed(digits);
+}
+
+// eslint-disable-next-line id-length
+function getRandomInt (a, b) {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+const getRandomElement = (items) => items[getRandomInt(0, items.length - 1)];
+
+export {getRandomInt, getRandomFloat, getRandomElement};
